+++ supertokens_python/framework/fastapi/fastapi_request.py
@@ -60,7 +60,9 @@ class FastApiRequest(BaseRequest):
         self.request.state.supertokens = None
 
     def get_path(self) -> str:
-        return self.request.url.path
+        root_path = self.request.scope.get("root_path")
+        url = self.request.url.path
+        return url[url.startswith(root_path) and len(root_path) :]
 
     async def form_data(self):
         return dict(parse_qsl((await self.request.body()).decode("utf-8")))